cat: nbLine from int to long.
(nbLine from int to long, the epic adventure.)

* src/dctc/java/com/dataiku/dctc/command/cat/HeadCatPrinter.java: Need
  to pass to long. When we need to skip we use long (file size are
  stored on long.)

+++ src/dctc/java/com/dataiku/dctc/command/cat/LastestLineCatAlgorithm.java
@@ -19,16 +19,16 @@ class LastestLineCatAlgorithm extends AbstractCatAlgorithm {
         }
     }
 
-    public int getNbLine() {
+    public long getNbLine() {
         return nbLine;
     }
-    public void setNbLine(int nbLine) {
+    public void setNbLine(long nbLine) {
         this.nbLine = nbLine;
     }
-    public LastestLineCatAlgorithm withNbLine(int nbLine) {
+    public LastestLineCatAlgorithm withNbLine(long nbLine) {
         setNbLine(nbLine);
         return this;
     }
 
-    private int nbLine;
+    private long nbLine;
 }
